Reorganize posixlib unit-tests to follow current javalib practice
