+++ unit-tests/native/src/test/scala/org/scalanative/testsuite/posixlib/MonetaryTest.scala
@@ -1,4 +1,4 @@
-package posixlib
+package org.scalanative.testsuite.posixlib
 
 import org.junit.Test
 import org.junit.Assert._
