+++ unit-tests/native/src/test/scala/org/scalanative/testsuite/posixlib/netinet/InTest.scala
@@ -1,4 +1,4 @@
-package scala.scalanative.posix
+package org.scalanative.testsuite.posixlib
 package netinet
 
 import scalanative.posix.netinet.in._
