+++ unit-tests/native/src/test/scala/org/scalanative/testsuite/posixlib/sys/ResourceTest.scala
@@ -1,18 +1,18 @@
-package scala.scalanative
-package posix
+package org.scalanative.testsuite.posixlib
 package sys
 
 import org.junit.Test
 import org.junit.Assert._
 
-import scalanative.posix.errno._
 import scalanative.runtime.Platform
 import scalanative.meta.LinktimeInfo.isWindows
 import scalanative.unsafe.{CInt, Ptr, Zone, alloc}
 import scalanative.unsigned._
 
-import resource._, resourceOps._
-import timeOps._
+import scalanative.posix.errno._
+import scalanative.posix.sys.resource._
+import scalanative.posix.sys.resourceOps._
+import scalanative.posix.sys.timeOps._
 
 // Design notes:
 //
